sync-files.py: check if files are in git
When newly adding files to the identical file configuration, the script
will successfully check whether the file contents are equal.

However if one forgot to add the file to git, the script will happily
report a success, which is not ideal if it is used as a pre-commit hook.

This adds such a check.